rust: cargo fuzz + corrupted varint handling
This adds cargo fuzz, a libfuzzer-based fuzz testing tool, to the Rust
implementation.

Uncovered in the process of fuzz testing was a class of inputs which would
decode as varints even though they shouldn't (i.e. garbage inputs that would
decode in a non-roundtrippable manner, because the length prefix encoded an
integer with e.g. a large zero component)

A check has been added for this class of inputs in the Rust implementation. In a
subsequent commit these cases will be added to the test vectors, and the same
check added to other language implementations.