Revamp wire type ordering and field ID offsets
This commit moves the "message" and "sequence" wire types to the highest
numbered wire type identifiers (6 and 7).

It also changes the offset for the field ID from 3-bits to 4-bits. This
is to make room for a "critical bit" similar to the one used to identify
critical X.509 extensions.

Implementing basic support for the critical bit will happen in a
follow-up PR.