rust: remove `tai64`
The goal of putting `tai64` in here was to bake it into Veriform as a
primitive type. However it seems like a TAI64N-like encoding in Veriform
can be modeled as a "well-known type" instead, and the wire type space
is "full".

So, probably best to leave it as is to avoid confusing the history.