rust: initial Verihash PoC
This is a PoC implementation of Verihash. It is not yet fully correct
and does not properly handle "non-terminals" (nested messages and
sequences).

It's implemented as yet another handrolled state machine which processes
incoming decoding events and incrementally computes a hash. In the
future it might be nice to implement this in terms of something like
session types instead (however this commit is spiking out the initial
protocol, which is still incomplete).

Having the hasher integrated directly into the decoder event system will
eventually enable support for computing hashes of messages which contain
unknown fields, as the decoder will (eventually) continue processing
these fields anyway even if they weren't originally requested by the
`Message` impl.