rust: nested Verihash (part 2) - initial sequence hashing
This adds partial support for computing Verihash over sequences.

It's still riddled with TODOs and does not yet extract the resulting
digest of sequences and add it to the messages (or other sequences) that
contain them.

However, it does fully parameterize the sequence decoder with the digest
type, adds an initial `sequence::Hasher`, and lays the groundwork for
eventually committing to these hashes.

I'm committing it anyway despite its incomplete state because I'd like
to do some refactoring on Verihash computation before completing this
work and wanted to break it down into incremental steps (so expect a
"part 3" which completes the implementation)