rust: support caching computed Verihash message digests
Previously the `message::Decoder::finish_digest` method had the nice
property of consuming the decoder (and therefore the message hasher)
upon completion of a message. This previously ensured messages are only
digested once they have been fully processed by making any subsequent
processing of the same message unrepresentable once the digest has been
computed.

Unfortunately, in order to embed the computed digest into a struct, we
need to access it multiple times, so this commit changes the method name
and signature to `compute_digest(&mut self)` in order to allow it to be
accessed several times:

- Once to embed it in the message struct itself
- Once when including the digest of a nested message in the "Merkelized"
  Verihash computation of its parent

(more accesses than twice are presently allowed, but unnecessary)

With this commit, it's now possible to compute a digest of a message at
decoding time and store the digest alongside the contents of a message
in a struct.

This is, after all, the whole point: we can compute these digests for
arbitrary parts of a message, allowing signers to sign
subsections/nested messages, and include their signatures alongside the
signed portion in the parent message.

This commit also adds a `fill_digest()` method on the public `Decoder`
API which is a sort of hack for avoiding `GenericArray` and can be used
inside a `Message` implementation to consume the digest and store it in
a regular `[u8; _]`.

The intent is to effectively standardize on SHA-256 and always use a
`[u8; 32]` for this value, with a `#[digest(sha256)]` annotation in a
proc macro to request the hash.