rust: initial custom derive support for sequences
The initial custom derive PR (#131) did not add support for fields
containing sequences, and had `todo!()` instead.

This commit adds support for deriving `Message` on structs whose fields
are sequences.

This support is limited to sequences of messages. A further PR is needed
to generalize it over other wire types.