rust: custom derive support for #[digest(...)] fields
Adds initial support for a field in digest structs where the computed
Verihash digest of a parsed message is stored.

This commit also includes quite a bit of incidental refactoring which
adds `DeriveEnum` and `DeriveStruct` structs which own several
`TokenStream`s for the derived code in-progress, which makes it a bit
easier to manage which functions are associated with which custom derive
operations.